Ignore mousedown event if emitted on scrollbar
In my usage of touch-scroll where I pan a giant table's content, I chose not hide the scrollbars as they still serve their purpose for concise navigation and serving as indicators as to how much more data there is to view.

While testing in IE I noticed that the scrollbar was emitting mousedown events to touch-scroll's mousedown handler and thus you couldn't navigate using the scrollbar.  While not a problem when the scrollbar is hidden, it obviously becomes an issue when it is not.  

I don't know if you would consider this out of scope because you hide the scrollbar by default, but this fix should help correct the expected behavior and continue working well with others that didn't exhibit it.

Thanks!